+++ src/numerospseudoaleatorios/modelo/Fibonacci.java
@@ -68,11 +68,11 @@ public class Fibonacci implements Generador{
     }
 
     @Override
-    public List<Integer> generarAlt(Integer cantidad) {
-         List<Integer> numeros = generar(cantidad);
-        List<Integer> lista = new ArrayList<>();
+    public List<Double> generarAlt(Integer cantidad) {
+        List<Integer> numeros = generar(cantidad);
+        List<Double> lista = new ArrayList<>();
         for (int i = 0; i < cantidad; i++) {
-            lista.add(numeros.get(i)/this.m);
+            lista.add((double)numeros.get(i)/this.m);
         }
         return lista;
     }
